Save, Cancle 기능 추가, ViewController 추상화
